tests/wpt/mozilla/tests/mozilla/document_defaultView.html

<html>
    <head>
        <script src="/resources/testharness.js"></script>
        <script src="/resources/testharnessreport.js"></script>
    </head>
    <body>
        <script>
          test(function() {
            assert_equals(document.defaultView, window);
          });
        </script>
    </body>
</html>
